+++ ui/ducks/bridge/dummy-quotes.ts
